phase11-wave-J: closeout — all 12 gates pass, zero clang-tidy errors
Wave J closeout:
- Fix W5 gate: add auto_only/with_auto feed patterns in ms_selection.cpp
- Fix W6 gate: extract int_expr_to_taql/bound_to_taql helpers, refactor
  to_taql_where() to use them
- Make W8-W12 gate scripts executable
- Fix all 163 clang-tidy errors across the project (implicit bool
  conversions, value-param by-copy, branch clones, integer division,
  incorrect rounding, vector reserve, identifier naming)
- All 12 wave gates (W1-W12) pass
- All 102 tests pass
- Clean clang-tidy build with -Werror succeeds

+++ tests/image_regrid_test.cpp
@@ -81,8 +81,8 @@ static void test_regrid_doubled_resolution() {
     for (std::int64_t dy = 0; dy < 16; ++dy) {
         for (std::int64_t dx = 0; dx < 16; ++dx) {
             // dst world = (dx, dy); src pixel = world / 2 = (dx/2, dy/2).
-            auto sx = static_cast<std::int64_t>(std::round(dx / 2.0));
-            auto sy = static_cast<std::int64_t>(std::round(dy / 2.0));
+            auto sx = static_cast<std::int64_t>(std::round(static_cast<double>(dx) / 2.0));
+            auto sy = static_cast<std::int64_t>(std::round(static_cast<double>(dy) / 2.0));
 
             float expected = 0.0F;
             if (sx >= 0 && sx < 8 && sy >= 0 && sy < 8) {
@@ -193,7 +193,7 @@ static void test_regrid_out_of_bounds_zero_fill() {
 // ---------------------------------------------------------------------------
 // Main
 // ---------------------------------------------------------------------------
-int main() {
+int main() { // NOLINT(bugprone-exception-escape)
     std::cout << "image_regrid_test\n";
 
     test_regrid_doubled_resolution();
